Added a (slow) test to catch regressions in to_c32.
In preparation for SF ticket 2496513 I've added a test that
simply compares a pre-refactor dump of Joe Smith's to_c32 output
to Joe Smith's current to_c32 output. It's not part of the spec
suite because it's too slow (due to the loading of the snomed_problems
fixture) and is intended to be used for the to_c32 refactor.